Move orderbook fetching to src folder and expose it
