+++ scripts/stablex/transitive_orderbook.js
@@ -1,5 +1,5 @@
 const BatchExchangeViewer = artifacts.require("BatchExchangeViewer")
-const { getOpenOrdersPaginated } = require("./utilities.js")
+const { getOpenOrdersPaginated } = require("../../src/onchain_reading.js")
 const BN = require("bn.js")
 
 const { Orderbook, Offer, transitiveOrderbook } = require("../../typescript/common/orderbook.js")
@@ -41,7 +41,11 @@ const addItemToOrderbooks = function (orderbooks, item) {
 }
 
 const getAllOrderbooks = async function (instance, pageSize) {
-  const elements = await getOpenOrdersPaginated(instance, pageSize)
+  let elements = []
+  for await (const page of getOpenOrdersPaginated(instance.contract, pageSize)) {
+    console.log("Fetched Page")
+    elements = elements.concat(page)
+  }
   const orderbooks = new Map()
   elements.forEach((item) => {
     addItemToOrderbooks(orderbooks, item)
