+++ tsconfig.json
@@ -53,7 +53,7 @@
     /* Source Map Options */
     // "sourceRoot": "",                      /* Specify the location where debugger should locate TypeScript files instead of source locations. */
     // "mapRoot": "",                         /* Specify the location where debugger should locate map files instead of generated locations. */
-    "inlineSourceMap": false,                 /* Emit a single file with source maps instead of having a separate file. */
+    "inlineSourceMap": false /* Emit a single file with source maps instead of having a separate file. */,
     // "inlineSources": true,                 /* Emit the source alongside the sourcemaps within a single file; requires '--inlineSourceMap' or '--sourceMap' to be set. */
 
     /* Experimental Options */
@@ -63,5 +63,5 @@
     /* Advanced Options */
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
-  "include": ["src/**/*.ts", "./node_modules/bn.js-typings/index.d.ts"]
+  "files": ["src/orderbook.ts", "src/fraction.ts"]
 }
